Supports HDR tone-mapping;
Updates `Color` to hold floating-point channel values; adds `Vec3::ones()` constructor; adds `Buffer2D::get_mut()` method; updates `Pipeline::deferred_framebuffer` to hold `Vec3` color samples; updates default fragment shader to perform exposure-based HDR tone-mapping;

+++ src/effects/invert_effect.rs
@@ -14,7 +14,11 @@ impl Effect for InvertEffect {
             for x in 0..buffer.width {
                 let color = Color::from_u32(*buffer.get(x, y));
 
-                let inverse = Color::rgb(255 - color.r, 255 - color.g, 255 - color.b);
+                let inverse = Color::rgb(
+                    255 - color.r as u8,
+                    255 - color.g as u8,
+                    255 - color.b as u8,
+                );
 
                 buffer.set(x, y, inverse.to_u32());
             }
